refactor: reorganize popup and background, fix import paths
- Move popup.ts to popup/index.ts with updated imports
- Move background.ts to background/index.ts
- Update all service imports to use new directory structure
- Fix smart-assistant and ml-worker import paths to shared/types
